+++ src/index.js
@@ -1,5 +1,7 @@
 export {default as Money} from './Money';
 export {default as Currency} from './Currency';
+export {default as CurrencyStore} from './CurrencyStore';
 export {default as WealthError} from './errors/WealthError';
 export {default as CurrencyMismatchError} from './errors/CurrencyMismatchError';
+export {default as InvalidCurrencyError} from './errors/InvalidCurrencyError';
 export {default as WrongInputError} from './errors/WrongInputError';