+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.0] - 2018-02-11
+### Fixed
+- Currency mismatch bug
+
+### Added
+- Custom errors
+- Extra APIs for Money and Currency
+- Formatter
+
+### Removed
+- **BREAKING:** Currency#getSymbolOnLeft()
+- **BREAKING:** Currency#getSpaceBetweenAmountAndSymbol()
+- **BREAKING:** Currency#unformat()
+
 ## [0.0.1] - 2017-12-14
 ### Fixed
 - Wrong import statements in the documentation
@@ -18,4 +32,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release
 
+[0.1.0]: https://github.com/amirmohsen/wealth/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/amirmohsen/wealth/compare/v0.0.0...v0.0.1